+++ packages/analyzer/fixtures/07-plugin-lit/05-mixin-method-deny-list/package/my-element.js
@@ -1,0 +1,5 @@
+export const MyMixin = (superclass) => class extends superclass {
+  foo() {}
+  requestUpdate() {}
+  firstUpdated() {}
+}