+++ packages/analyzer/src/features/framework-plugins/lit/method-denylist.js
@@ -4,18 +4,17 @@
  * Excludes methods from the manifest
  */
 export function methodDenyListPlugin() {
-  const METHOD_DENY_LIST = ['requestUpdate', 'performUpdate', 'shouldUpdate', 'update', 'render', 'firstUpdated', 'updated', 'willUpdate'];
+  const METHOD_DENY_LIST = ['requestUpdate', 'createRenderRoot', 'scheduleUpdate', 'performUpdate', 'shouldUpdate', 'update', 'render', 'firstUpdated', 'updated', 'willUpdate'];
 
   return {
     name: 'CORE - LIT-METHOD-DENYLIST',
     moduleLinkPhase({moduleDoc}){
-      const classes = moduleDoc?.declarations?.filter(declaration => declaration.kind === 'class');
+      const classesAndMixins = moduleDoc?.declarations?.filter(declaration => declaration.kind === 'class' || declaration.kind === 'mixin');
 
-      classes?.forEach(klass => {
+      classesAndMixins?.forEach(klass => {
         if(!klass?.members) return;
         klass.members = klass?.members?.filter(member => !METHOD_DENY_LIST.includes(member.name));
       });
     },
   }
 }
-